Consolidate LLM metrics module
